The product addition section has been completed.
The interface has been maximized for full-screen operation. Products can now be added from the interface. A SQL connection has been established to the database, and once all product information is entered and the save button is clicked, the product is added. If there are empty fields, the product is not added.